각종 함수들 레이어에 맞게 분리
