+++ src/basic/constants/productConstants.js
@@ -1,0 +1,6 @@
+// 상품 식별자 상수
+export const PRODUCT_ONE = "p1";
+export const PRODUCT_TWO = "p2";
+export const PRODUCT_THREE = "p3";
+export const PRODUCT_FOUR = "p4";
+export const PRODUCT_FIVE = "p5";
