+++ src/basic/domains/index.js
@@ -1,0 +1,3 @@
+export * from "./points.js";
+export * from "./product.js";
+export * from "./cart.js";
