+++ src/basic/state/appState.js
@@ -1,0 +1,11 @@
+// 애플리케이션 전역 상태
+export let appState = {
+  prodList: [],
+  bonusPts: 0,
+  itemCnt: 0,
+  lastSel: null,
+  totalAmt: 0,
+};
+
+// DOM 요소 참조 저장소
+export let domElements = {};
